Separated out the table entry edit modal stuff into its own file, since it was getting long.
